pyproject.toml
[project]
name = "sofia-agent"
version = "0.1.9"
description = "SOFIA: Simple Orchestrated Flow Intelligence Agent"
authors = [
    {name = "chandralegend",email = "irugalbandarachandra@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "loguru (>=0.7.3,<0.8.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "docstring-parser (>=0.16,<0.17)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project.optional-dependencies]
cli = [
    "questionary (>=2.1.0,<3.0.0)",
]
openai = [
    "openai (>=1.76.0,<2.0.0)",
]
mistralai = [
    "mistralai (>=1.7.0,<2.0.0)",
]
gemini = [
    "google-genai (>=1.12.1,<2.0.0)",
]

[project.scripts]
sofia = "sofia_agent.cli:main"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=sofia_agent --cov-report=term-missing"
testpaths = [
    "tests",
]


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.1.1"
pytest-asyncio = "^0.26.0"
sqlmodel = "^0.0.24"
redis = "^6.0.0"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"
pre-commit = "^4.2.0"
